+++ src/core/runtime/conversions.js
@@ -52,25 +52,10 @@ export const conversions = {
         },
         // Values conversion - extracts form values from DOM nodes
         function(str, node, runtime) {
-            if (!(str === "Values" || str.startsWith("Values:"))) {
-                return;
-            }
-            var conversion = str.split(":")[1];
+            if (str !== "Values") return;
             var formData = new HyperscriptFormData();
-
             runtime.implicitLoop(node, (node) => formData.addContainer(node));
-
-            if (conversion) {
-                if (conversion === "JSON") {
-                    return JSON.stringify(formData.result);
-                } else if (conversion === "Form") {
-                    return new URLSearchParams(formData.result).toString();
-                } else {
-                    throw new Error("Unknown conversion: " + conversion);
-                }
-            } else {
-                return formData.result;
-            }
+            return formData.result;
         },
     ],
     String: function (val) {
@@ -99,6 +84,9 @@ export const conversions = {
         return Array.from(val);
     },
     JSON: function (val) {
+        return JSON.parse(val);
+    },
+    JSONString: function (val) {
         return JSON.stringify(val);
     },
     Object: function (val) {
@@ -110,6 +98,9 @@ export const conversions = {
         } else {
             return Object.assign({}, val);
         }
+    },
+    FormEncoded: function (val) {
+        return new URLSearchParams(val).toString();
     },
     HTML: function (value) {
         var toHTML = (value) => {